Remove sublocations necessary for area constraints
So far, there were 4 sublocations per node, so 5 nodes per location in total. This is bad for performance. The sublocations existed so that technologies can compete on the same land. But actually, only open field pv and (parts of) wind onshore are competing on the same land. So now, all other renewable technologies are restricted by power caps. The area of a location represents the land that can be used by open field pv and wind onshore. This required a second wind onshore technology, as one has a area footprint (the one competing with open field pv) whereas the other does not.

The whole change removes the need for sublocations and thus removes 80% of the Calliope locations. In one sample, this removed the runtime by about 40%.